backend/requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# AI & ML
anthropic>=0.34.0
openai>=1.12.0
transformers>=4.35.0
torch>=2.0.0
sentencepiece>=0.1.99
protobuf>=4.25.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
audioread>=3.0.0
pydub>=0.25.0
scipy>=1.11.0
numba>=0.57.0

# Image Processing & OCR
Pillow>=10.0.0
pytesseract>=0.3.10
opencv-python>=4.8.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Utilities
python-dateutil>=2.8.0
requests>=2.31.0
aiofiles>=23.2.0
